NoneDiscountPolicy의 개념적 혼란을 제거하라
* `DiscountPolicy`를 `DefaultDiscountPolicy`로 개명했습니다.
    * `PercentDiscountPolicy`는 `DefaultDiscountPolicy`를 상속합니다.
    * `AmountDiscountPolicy`는 `DefaultDiscountPolicy`를 상속합니다.
* `DiscountPolicy` 인터페이스를 새로 추가했습니다.
    * `NoneDiscountPolicy`는 `DiscountPolicy`를 구현합니다.
