+++ src/main/java/com/johngrib/objects/_02_movie/DefaultDiscountPolicy.java
@@ -1,0 +1,35 @@
+package com.johngrib.objects._02_movie;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+
+/** 기본 할인 정책. */
+public abstract class DefaultDiscountPolicy implements DiscountPolicy {
+  /** 하나의 할인 정책은 여러 개의 할인 조건을 포함할 수 있다. */
+  private List<DiscountCondition> conditions = new ArrayList<>();
+
+  public DefaultDiscountPolicy(DiscountCondition... conditions) {
+    this.conditions = Arrays.asList(conditions);
+  }
+
+  /**
+   * 할인 금액을 계산한다.
+   * <p>
+   * Template Method.
+   *
+   * @param screening
+   * @return
+   */
+  @Override
+  public Money calculateDiscountAmount(Screening screening) {
+    for (DiscountCondition each : conditions) {
+      if (each.isSatisfiedBy(screening)) {
+        return getDiscountAmount(screening);
+      }
+    }
+    return Money.ZERO;
+  }
+
+  abstract protected Money getDiscountAmount(Screening screening);
+}
