+++ src/main/java/com/johngrib/objects/_02_movie/DiscountPolicy.java
@@ -1,34 +1,12 @@
 package com.johngrib.objects._02_movie;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
 /** 할인 정책. */
-public abstract class DiscountPolicy {
-  /** 하나의 할인 정책은 여러 개의 할인 조건을 포함할 수 있다. */
-  private List<DiscountCondition> conditions = new ArrayList<>();
-
-  public DiscountPolicy(DiscountCondition... conditions) {
-    this.conditions = Arrays.asList(conditions);
-  }
-
+public interface DiscountPolicy {
   /**
    * 할인 금액을 계산한다.
-   * <p>
-   * Template Method.
    *
-   * @param screening
-   * @return
+   * @param screening 상영 정보
+   * @return 할인 금액
    */
-  public Money calculateDiscountAmount(Screening screening) {
-    for (DiscountCondition each : conditions) {
-      if (each.isSatisfiedBy(screening)) {
-        return getDiscountAmount(screening);
-      }
-    }
-    return Money.ZERO;
-  }
-
-  abstract protected Money getDiscountAmount(Screening screening);
+  Money calculateDiscountAmount(Screening screening);
 }
