+++ src/main/java/com/johngrib/objects/_02_movie/NoneDiscountPolicy.java
@@ -3,9 +3,9 @@ package com.johngrib.objects._02_movie;
 /**
  * 할인하지 않는 할인 정책.
  */
-public class NoneDiscountPolicy extends DiscountPolicy {
+public class NoneDiscountPolicy implements DiscountPolicy {
   @Override
-  protected Money getDiscountAmount(Screening screening) {
+  public Money calculateDiscountAmount(Screening screening) {
     return Money.ZERO;
   }
 }
