+++ src/main/java/com/johngrib/objects/_02_movie/PercentDiscountPolicy.java
@@ -1,7 +1,7 @@
 package com.johngrib.objects._02_movie;
 
 /** 일정 비율 할인 정책. */
-public class PercentDiscountPolicy extends DiscountPolicy {
+public class PercentDiscountPolicy extends DefaultDiscountPolicy {
   /** 할인 비율. */
   private double percent;
